requirements.txt
pip~=24.2
attrs~=24.2.0
django-request~=1.6.3
filelock~=3.14.0
redis~=5.1.1
setuptools~=69.5.1
packaging~=24.0
platformdirs~=4.2.2
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1